+++ db/migrations/0016_create_vocabulary_entries.sql
@@ -7,7 +7,7 @@ SET @sql := IF(@exist = 0,
   'CREATE TABLE vocabulary_entries (
     vocabulary_entry_id INT AUTO_INCREMENT PRIMARY KEY,
     child_profile_id INT NOT NULL,
-    story_id INT NULL,
+    story_id BIGINT NULL,
     english_word VARCHAR(100) NOT NULL,
     korean_meaning VARCHAR(255) NOT NULL,
     example_sentence TEXT NULL,
